==> 2_Metabolomics/bin/Rstudio/1_2.R <==
### This script make a dataframe with all samples of GC-SM analisys###
## February 2020
## Verónica Reyes Galindo

#Load libraries
library (ggplot2)
library("devtools")

#Load files
all_metabolites <- read.delim("../../metadata/htm_df_allsamples.txt")

# Remove data with Qual <80
a_metabol_Q80 <- subset(all_metabolites, Qual > 80, select = c("Id","Pk","RT","Area", "Ref", "CAS", "Qual"))

#Change Name CAS
beta <- intToUtf8(946)
alpha <- intToUtf8(945)
delta <- intToUtf8(948)
gamma <- intToUtf8(947)
a_metabol_Q80$CAS <- as.character(a_metabol_Q80$CAS)


a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000127-91-3"]<- paste0(beta,"-Pinene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "018172-67-3"]<- paste0( "L-",beta,"-Pinene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "007785-70-8"]<- paste0("1R-",beta,"-Pinene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "003387-41-5"]<- paste0(beta,"-Sabinene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "013466-78-9"]<- paste0("3-",delta,"-Carene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000099-85-4"]<- paste0( gamma,"-Terpinen")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000138-86-3"]<- paste0(alpha,"-Limonene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "005989-27-5"]<- paste0( "D-Limonene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000498-15-7"]<- paste0("(+)-3-Carene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000586-62-9"]<- paste0(alpha,"-Terpinolene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000554-61-0"]<- paste0("2-Carene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "029714-87-2"]<- paste0("Ocimene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000099-89-8"]<- paste0( "Phenol, 4-(1-methylethyl)-")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000618-45-1"]<- paste0("3-Isopropylphenol")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000076-49-3"]<- paste0("Phenol, 3-(1-methylethyl)-")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "005655-61-8"]<- paste0("L-bornyl acetate")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "017699-14-8"]<- paste0(alpha,"-Cubebene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "092618-89-8"]<- paste0("Acetic acid, bornyl ester")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "003856-25-5"]<- paste0( alpha,"-Copaene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "013744-15-5"]<- paste0(beta,"-Cubebene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "150320-52-8"]<- paste0("Bicyclo[4.4.0]dec-1-ene, 2-isopropyl-5-methyl-9-methylene-")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000087-44-5"]<- paste0("Caryophyllene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000118-65-0"]<- paste0(beta,"-cis-Caryophyllene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "006753-98-6"]<- paste0("Humulene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "003338-55-4"]<- paste0(beta,"-cis-Ocimene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "031983-22-9"]<- paste0(alpha,"-Muurolene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000483-75-0"]<- paste0(alpha,"-Amorphene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "030021-74-0"]<- paste0(gamma,"-Muurolene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000483-76-1"]<- paste0(delta,"-Cadinene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "054324-03-7"]<- paste0("Bicyclosesquiphellandrene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "039029-41-9"]<- paste0(gamma,"-Cadinene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "016728-99-7"]<- paste0("Cadinadiene-1,4")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000469-61-4"]<- paste0(alpha,"-Cedrene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "005951-67-7"]<- paste0(alpha,"-Elemene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000075-09-2"]<- paste0("Methylene chloride")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "013877-91-3"]<- paste0(beta,"-Ocimene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "005208-50-4"]<- paste0("Cadinadiene-1,4")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "002437-95-8"]<- paste0(alpha,"-Pinene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "003779-61-1"]<- paste0("trans-",beta,"-Ocimene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000099-84-3"]<- paste0(beta,"-Terpinene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "029548-02-5"]<- paste0("Artemisia triene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "007785-26-4"]<- paste0("L-",alpha,"-Pinene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "007705-14-8"]<- paste0(alpha,"-Limonene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000123-35-3"]<- paste0(beta,"-Myrcene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "028935-76-4"]<- paste0("trans-",beta,"-Ocimene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "060909-27-5"]<- paste0("Himachala-2,4-diene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "022567-17-5"]<- paste0(gamma,"-Gurjunene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "025246-27-9"]<- paste0("Alloaromadendrene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "041702-63-0"]<- paste0("Epizonarene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "001405-16-9"]<- paste0("Patchoulene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "016609-28-2"]<- paste0("Cyclohexene, 1,5,5-trimethyl-3-methylene-")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "023986-74-5"]<- paste0("Germacrene D")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "006813-21-4"]<- paste0("Selina-3,7(11)-diene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "020071-49-2"]<- paste0(gamma,"-Maaliene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "022567-17-5"]<- paste0(gamma,"-Gurjunene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000489-29-2"]<- paste0(beta,"-Maaliene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "001686-67-5"]<- paste0("Rimuene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000150-86-7"]<- paste0("Phytol")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "023283-97-8"]<- paste0("(1S,2R,5R)-(+)-Isomenthol")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "010214-27-4"]<- paste0("iso-Amyl 10-undecenoate")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000511-15-9"]<- paste0("Totarol")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "001873-88-7"]<- paste0("Bis(trimethylsiloxy)methylsilane")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000107-86-8"]<- paste0(beta,"-Methylcrotonaldehyde")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000075-00-3"]<- paste0("Ethyl Chloride")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "034314-83-5"]<- paste0("4-Methyl-2,3-dihydrofuran")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "0000099-86-5"]<- paste0(alpha,"-Terpinene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000125-12-2"]<- paste0("Isobornyl acetate")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "017334-55-3"]<- paste0(beta,"-Gurjunene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "002958-62-5"]<- paste0(beta,"-Methylcrotonaldehyde")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000062-61-9"]<- paste0("Ethyl Chloride")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "010208-80-7"]<- paste0("4-Methyl-2,3-dihydrofuran")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000125-12-2"]<- paste0("Isobornyl acetate")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "017334-55-3"]<- paste0(beta,"-Gurjunene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "017334-55-3"]<- paste0("Bis(trimethylsiloxy)methylsilane")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "0000099-86-5"]<- paste0(alpha,"-Terpinene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000079-92-5"]<- paste0("Camphene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000059-02-9"]<- paste0("Vitamin E")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000080-56-8"]<- paste0(alpha,"-Pinene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000088-69-7"]<- paste0("Phenol")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000481-34-5"]<- paste0(alpha,"-Cadinol")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000502-99-8"]<- paste0(alpha,"-Ocimene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "001687-64-5"]<- paste0("o-Cresol")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "001235-74-1"]<- paste0("Methyl dehydroabietate")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "003513-69-7"]<- paste0("Methyl levopimarate")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "001139-30-6"]<- paste0(beta,"-Caryophyllene oxide")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "001200-67-5"]<- paste0("Isobornyl formate")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "003760-14-3"]<- paste0("1,5-Cyclooctadiene, 1,5-dimethyl-")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "005989-54-8"]<- paste0("L-Limonene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000062-28-4"]<- paste0("1000062-28-4")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "339239-69-9"]<- paste0("1000340-19-1")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "053282-47-6"]<- paste0("053282-47-6")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000818-48-4"]<- paste0("000818-48-4")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000145-84-2"]<- paste0("1000145-84-2")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000062-61-9"]<- paste0("1000062-61-9")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000499-03-6"]<- paste0("000499-03-6")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000185-13-4"]<- paste0("1000185-13-4")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000210-81-9"]<- paste0("1000210-81-9")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "061477-41-6"]<- paste0("061477-41-6")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "072120-50-4"]<- paste0("072120-50-4")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "067843-59-8"]<- paste0("067843-59-8")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "072120-50-4"]<- paste0("072120-50-4")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000159-39-0"]<- paste0("1000159-39-0")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000326-16-3"]<- paste0("1000326-16-3")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "017367-38-3"]<- paste0("017367-38-3")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000197-42-9"]<- paste0("1000197-42-9")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "094427-47-1"]<- paste0("094427-47-1")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "024035-43-6"]<- paste0("024035-43-6")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "002768-84-5"]<- paste0("002768-84-5")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000197-90-8"]<- paste0("1000197-90-8")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000340-19-1"]<- paste0("1000340-19-1")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000818-48-4"]<- paste0("000818-48-4")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000145-84-2"]<- paste0("1000145-84-2")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000156-10-8"]<- paste0("1000156-10-8")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000192-43-5"]<- paste0("1000192-43-5")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "072120-50-4"]<- paste0("072120-50-4")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "004889-83-2"]<- paste0("004889-83-2")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1068998-21-0"]<- paste0("1068998-21-0")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "017627-24-6"]<- paste0("017627-24-6")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "005951-61-1"]<- paste0("005951-61-1")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "005794-04-7"]<- paste0("005794-04-7")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "124790-36-9"]<- paste0("124790-36-9")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000162-76-8"]<- paste0("1000162-76-8")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000162-25-4"]<- paste0("1000162-25-4")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000159-39-0"]<- paste0("1000159-39-0")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000156-12-4"]<- paste0("1000156-12-4")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000151-99-7"]<- paste0("1000151-99-7")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000150-36-1"]<- paste0("1000150-36-1")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000145-84-3"]<- paste0("1000145-84-3")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000145-84-2"]<- paste0("1000145-84-2")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "1000140-07-4"]<- paste0("1000140-07-4")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "056348-21-1"]<- paste0("056348-21-1")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000511-02-4"]<- paste0("Sclarene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000508-32-7"]<- paste0("Tricyclene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "056710-83-9"]<- paste0("Cyclopentane, 2-methyl-1-methylene-3-(1-methylethenyl)-")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "055255-51-1"]<- paste0("055255-51-1")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "051898-92-1"]<- paste0("Tetracyclo[6.1.0.0(2,4).0(5,7)]nonane,3,3,6,6,9,9-hexamethyl-")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "050894-66-1"]<- paste0(alpha,"-Funebrene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000523-47-7"]<- paste0(beta,"-Cadinene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000515-17-3"]<- paste0(gamma,"-Selinene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000514-51-2"]<- paste0(beta,"-Patchoulene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000630-02-4"]<- paste0("Octacosane")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "028634-89-1"]<- paste0(beta,"-Thujene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000586-63-0"]<- paste0("Isoterpinolene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000555-10-2"]<- paste0(beta,"-Phellandrene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000552-02-3"]<- paste0("Viridiflorol")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000546-28-1"]<- paste0(beta,"-Cedrene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000586-81-2"]<- paste0(gamma,"-Terpineol")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "029050-33-7"]<- paste0(delta,"-Carene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "029837-07-8"]<- paste0("cis-",alpha,"-Bisabolene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "024406-05-1"]<- paste0(alpha,"-Cadinene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "022469-52-9"]<- paste0("Cyclosativene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "006630-01-9"]<- paste0("Benzene, 1-(1,1-dimethylethyl)-3-ethyl-5-methyl-")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "019912-62-0"]<- paste0(tau,"-Muurolol")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "017066-67-0"]<- paste0(beta,"-Selinene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "010219-75-7"]<- paste0("Eremophilene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000638-68-6"]<- paste0("Triacontane")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "003691-11-0"]<- paste0(delta,"-Guaijene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "013877-93-5"]<- paste0("Bicyclo[7.2.0]undec-4-ene, 4,11,11-trimethyl-8-methylene-")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "010191-41-0"]<- paste0("dl-α-Tocopherol")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000646-31-1"]<- paste0("Tetracosane")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "019402-34-7"]<- paste0("019402-34-7")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "001135-66-6"]<- paste0("Isolongipholene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "005937-11-1"]<- paste0(tau,"-Cadinol")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000489-40-7"]<- paste0(alpha,"-Gurjunene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000475-20-7"]<- paste0("Longifolene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000118-71-8"]<- paste0("Maltol")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000096-76-4"]<- paste0("2,4-Di-tert-butylphenol")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000464-17-5"]<- paste0("Bornylene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000473-14-3"]<- paste0("000473-14-3")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000473-13-2"]<- paste0(alpha,"-Selinene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000489-39-4"]<- paste0("Aromandendrene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000099-83-2"]<- paste0(alpha,"-Phellandrene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "005794-03-6"]<- paste0("Camphene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "001138-52-9"]<- paste0("Phenol, 3,5-di-tert-butyl-")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "005875-45-6"]<- paste0("Phenol, 2,5-di-tert-butyl-")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "004630-07-3"]<- paste0("Valencene")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000717-74-8"]<- paste0("Benzene, 1,3,5-triisopropyl-")
a_metabol_Q80$CAS[a_metabol_Q80$CAS == "000099-86-5"]<- paste0(alpha,"-Terpinene")
